+++ docs/quality/changelog.md
@@ -872,3 +872,8 @@
 - `2026-04-13 12:59:15` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/requirements.txt`
 - `2026-04-13 12:59:24` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/.env.example`
 - `2026-04-13 12:59:31` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/.gitignore`
+- `2026-04-13 13:02:41` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/main.py`
+- `2026-04-13 13:02:49` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/main.py`
+- `2026-04-13 13:02:58` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/main.py`
+- `2026-04-13 13:03:22` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_cli/test_main.py`
+- `2026-04-13 13:04:21` [feat/deploy-setup] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_cli/test_main.py`